add Socket Service Fabric
