+++ study-web-server/src/main/java/ru/geekbrains/RequestHandler.java
@@ -2,6 +2,7 @@ package ru.geekbrains;
 
 import ru.geekbrains.domain.HttpRequest;
 import ru.geekbrains.domain.HttpResponse;
+import ru.geekbrains.services.SocketService;
 
 import java.io.IOException;
 import java.util.Deque;
