+++ study-web-server/src/main/java/ru/geekbrains/services/SocketServiceFactory.java
@@ -1,0 +1,10 @@
+package ru.geekbrains.services;
+
+import java.net.Socket;
+
+public class SocketServiceFactory {
+
+    public static SocketService create(Socket socket) {
+        return new SocketServiceImpl(socket);
+    }
+}
